requirements.txt
asgiref==3.5.2
autopep8==2.0.0
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
django-filter==22.1
djangorestframework==3.14.0
fake-useragent==0.1.14
idna==3.4
importlib-metadata==5.0.0
Markdown==3.4.1
pycodestyle==2.9.1
pytz==2022.6
requests==2.28.1
soupsieve==2.3.2.post1
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.6
urllib3==1.26.12
zipp==3.10.0
